Separating set() and subset()
